+++ src/main/java/com/tnki/core/auth/service/impl/SecurityServiceImpl.java
@@ -57,7 +57,7 @@ public class SecurityServiceImpl implements SecurityService {
     }
 
     @Override
-    public int autoLogin(String username, String password) {
+    public long autoLogin(String username, String password) {
         UserDetails userDetails = userDetailsService.loadUserByUsername(username);
         UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(userDetails, password, userDetails.getAuthorities());
         authenticationManager.authenticate(usernamePasswordAuthenticationToken);
